requirements.txt
pyqt6>=6.6
sqlalchemy>=2.0
openpyxl>=3.1
pyyaml>=6.0
matplotlib>=3.8
pyinstaller>=6.0
pyarmor>=8.0
